requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
bidict==0.23.1
frozenlist==1.4.1
h11==0.14.0
idna==3.8
multidict==6.0.5
python-engineio==4.9.1
python-socketio==5.11.3
simple-websocket==1.0.0
wsproto==1.2.0
yarl==1.9.4
